Minor fixes

+++ system/session.js
@@ -84,7 +84,8 @@ module.exports = (EXPIRE_TIME = 3600, DOMAIN = null) => {
                  * Session save
                  */
                 req.session.Save = () => {
-                    fs.writeFileSync(P, JSON.stringify(Buffer));
+                    if (fs.existsSync(P))
+                        fs.writeFileSync(P, JSON.stringify(Buffer));
                 }
                 next();
           };
